backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.3
pyfhel==3.4.2
cryptography==42.0.8
python-multipart==0.0.9
slowapi==0.1.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
